+++ src/pytorch_lightning/core/optimizer.py
@@ -288,11 +288,6 @@ def _configure_schedulers_automatic_opt(schedulers: list, monitor: Optional[str]
                     "reduce_on_plateau",
                     isinstance(scheduler["scheduler"], optim.lr_scheduler.ReduceLROnPlateau)
             )
-            elif not isinstance(scheduler["reduce_on_plateau"], bool):
-                raise MisconfigurationException(
-                    'The "reduce_on_plateau" key in lr scheduler dict must be a bool'
-                    f' but is "{type(scheduler["reduce_on_plateau"])}"'
-                )
             if scheduler["reduce_on_plateau"] and scheduler.get("monitor", None) is None:
                 raise MisconfigurationException(
                     "The lr scheduler dict must include a monitor when a `ReduceLROnPlateau` scheduler is used."